to:小小细节

+++ cmd/api_gateway/api/comment.go
@@ -75,7 +75,7 @@ func PostCommentAction(c context.Context, ctx *app.RequestContext) {
 	commentId, _ := strconv.Atoi(ctx.Query("comment_id"))
 
 	req := &comment.CommentActionRequest{
-		UserId:      int64(userId),
+		UserId:      userId,
 		VideoId:     int64(videoId),
 		ActionType:  int32(actionType),
 		CommentText: commentText,
